Clean up tracing API.
