+++ Cesium3DTiles/src/GltfContent.cpp
@@ -3,7 +3,7 @@
 #include "CesiumGltf/AccessorView.h"
 #include "CesiumGltf/AccessorWriter.h"
 #include "CesiumUtility/Math.h"
-#include "CesiumUtility/Profiler.h"
+#include "CesiumUtility/Tracing.h"
 #include "CesiumUtility/joinToString.h"
 #include <stdexcept>
 
@@ -20,7 +20,7 @@ GltfContent::load(const TileContentLoadInput& input) {
     const std::shared_ptr<spdlog::logger>& pLogger,
     const std::string& url,
     const gsl::span<const std::byte>& data) {
-  TRACE("Cesium3DTiles::GltfContent::load")
+  CESIUM_TRACE("Cesium3DTiles::GltfContent::load")
   std::unique_ptr<TileContentLoadResult> pResult =
       std::make_unique<TileContentLoadResult>();
 
@@ -61,7 +61,7 @@ static int generateOverlayTextureCoordinates(
     double& north,
     double& minimumHeight,
     double& maximumHeight) {
-  TRACE("Cesium3DTiles::GltfContent::generateOverlayTextureCoordinates")
+  CESIUM_TRACE("Cesium3DTiles::GltfContent::generateOverlayTextureCoordinates")
   std::vector<CesiumGltf::Buffer>& buffers = gltf.buffers;
   std::vector<CesiumGltf::BufferView>& bufferViews = gltf.bufferViews;
   std::vector<CesiumGltf::Accessor>& accessors = gltf.accessors;
@@ -189,7 +189,8 @@ GltfContent::createRasterOverlayTextureCoordinates(
     uint32_t textureCoordinateID,
     const CesiumGeospatial::Projection& projection,
     const CesiumGeometry::Rectangle& rectangle) {
-  TRACE("Cesium3DTiles::GltfContent::createRasterOverlayTextureCoordinates")
+  CESIUM_TRACE(
+      "Cesium3DTiles::GltfContent::createRasterOverlayTextureCoordinates")
   std::vector<int> positionAccessorsToTextureCoordinateAccessor;
   positionAccessorsToTextureCoordinateAccessor.resize(gltf.accessors.size(), 0);
