+++ Cesium3DTiles/src/QuantizedMeshContent.cpp
@@ -6,7 +6,7 @@
 #include "CesiumGeospatial/GlobeRectangle.h"
 #include "CesiumUtility/JsonHelpers.h"
 #include "CesiumUtility/Math.h"
-#include "CesiumUtility/Profiler.h"
+#include "CesiumUtility/Tracing.h"
 #include "CesiumUtility/Uri.h"
 #include "SkirtMeshMetadata.h"
 #include "calcQuadtreeMaxGeometricError.h"
@@ -697,7 +697,7 @@ QuantizedMeshContent::load(const TileContentLoadInput& input) {
     const gsl::span<const std::byte>& data,
     bool enableWaterMask) {
 
-  TRACE("Cesium3DTiles::QuantizedMeshContent::load")
+  CESIUM_TRACE("Cesium3DTiles::QuantizedMeshContent::load")
 
   // TODO: use context plus tileID to compute the tile's rectangle, rather than
   // inferring it from the parent tile.
