+++ Cesium3DTiles/src/Tileset.cpp
@@ -14,7 +14,7 @@
 #include "CesiumGeospatial/GeographicProjection.h"
 #include "CesiumUtility/JsonHelpers.h"
 #include "CesiumUtility/Math.h"
-#include "CesiumUtility/Profiler.h"
+#include "CesiumUtility/Tracing.h"
 #include "CesiumUtility/Uri.h"
 #include "TileUtilities.h"
 #include "calcQuadtreeMaxGeometricError.h"
@@ -54,7 +54,7 @@ Tileset::Tileset(
       _tileDataBytes(0),
       _supportsRasterOverlays(false),
       _gltfUpAxis(CesiumGeometry::Axis::Y) {
-  TRACE_ASYNC_ENLIST(CesiumUtility::Profiler::instance().allocateID());
+  CESIUM_TRACE_ASYNC_ENLIST(CesiumUtility::Profiler::instance().allocateID());
 
   ++this->_loadsInProgress;
   this->_loadTilesetJson(url);
@@ -84,8 +84,8 @@ Tileset::Tileset(
       _overlays(*this),
       _tileDataBytes(0),
       _supportsRasterOverlays(false) {
-  TRACE_ASYNC_ENLIST(CesiumUtility::Profiler::instance().allocateID());
-  TRACE_ASYNC_BEGIN("Tileset from ion startup");
+  CESIUM_TRACE_ASYNC_ENLIST(CesiumUtility::Profiler::instance().allocateID());
+  CESIUM_TRACE_BEGIN("Tileset from ion startup");
 
   std::string ionUrl = "https://api.cesium.com/v1/assets/" +
                        std::to_string(ionAssetID) + "/endpoint";
@@ -108,7 +108,7 @@ Tileset::Tileset(
             e.what());
         this->notifyTileDoneLoading(nullptr, 0);
       })
-      .thenImmediately([]() { TRACE_ASYNC_END("Tileset from ion startup"); });
+      .thenImmediately([]() { CESIUM_TRACE_END("Tileset from ion startup"); });
 }
 
 Tileset::~Tileset() {
@@ -135,7 +135,7 @@ Tileset::~Tileset() {
 
   for (size_t i = 0; i < this->_loadingIDs.size(); ++i) {
     int64_t id = this->_loadingIDs[i];
-    TRACE_ASYNC_END_ID(
+    CESIUM_TRACE_END_ID(
         ("Tileset Loading Slot " + std::to_string(id)).c_str(),
         id);
   }
@@ -417,7 +417,7 @@ int64_t Tileset::notifyTileStartLoading(Tile* pTile) noexcept {
       int64_t loaderIndex = it - this->_tilesBeingLoaded.begin();
       loaderID = this->_loadingIDs[loaderIndex];
 
-      TRACE_ASYNC_BEGIN_ID(
+      CESIUM_TRACE_BEGIN_ID(
           TileIdUtilities::createTileIdString(pTile->getTileID()).c_str(),
           loaderID);
     } else {
@@ -444,7 +444,7 @@ void Tileset::notifyTileDoneLoading(
     if (it != this->_tilesBeingLoaded.end()) {
       *it = nullptr;
 
-      TRACE_ASYNC_END_ID(
+      CESIUM_TRACE_END_ID(
           TileIdUtilities::createTileIdString(pTile->getTileID()).c_str(),
           loaderID);
     } else {
@@ -483,7 +483,7 @@ Tileset::RequestTileContentResult Tileset::requestTileContent(Tile& tile) {
 
   int64_t loaderID = this->notifyTileStartLoading(&tile);
 
-  TRACE_ASYNC_ENLIST(loaderID);
+  CESIUM_TRACE_ASYNC_ENLIST(loaderID);
 
   return RequestTileContentResult{
       this->getExternals().pAssetAccessor->requestAsset(
@@ -529,7 +529,7 @@ Future<void> Tileset::_loadTilesetJson(
   }
   pContext->pTileset = this;
 
-  TRACE_ASYNC_BEGIN("Load tileset.json");
+  CESIUM_TRACE_BEGIN("Load tileset.json");
 
   return this->getExternals()
       .pAssetAccessor->requestAsset(this->getAsyncSystem(), url, headers)
@@ -549,7 +549,7 @@ Future<void> Tileset::_loadTilesetJson(
         this->addContext(std::move(loadResult.pContext));
         this->_pRootTile = std::move(loadResult.pRootTile);
         this->notifyTileDoneLoading(nullptr, 0);
-        TRACE_ASYNC_END("Load tileset.json");
+        CESIUM_TRACE_END("Load tileset.json");
       })
       .catchInMainThread([this, url](const std::exception& e) {
         SPDLOG_LOGGER_ERROR(
@@ -559,7 +559,7 @@ Future<void> Tileset::_loadTilesetJson(
             e.what());
         this->_pRootTile.reset();
         this->notifyTileDoneLoading(nullptr, 0);
-        TRACE_ASYNC_END("Load tileset.json");
+        CESIUM_TRACE_END("Load tileset.json");
       });
 }
 
@@ -1790,7 +1790,7 @@ void Tileset::_processLoadQueue() {
     for (size_t i = 0; i < this->_loadingIDs.size(); ++i) {
       int64_t id = Profiler::instance().allocateID();
       this->_loadingIDs[i] = id;
-      TRACE_ASYNC_BEGIN_ID(
+      CESIUM_TRACE_BEGIN_ID(
           ("Tileset Loading Slot " + std::to_string(id)).c_str(),
           id);
     }
