+++ CesiumGeometry/src/clipTriangleAtAxisAlignedThreshold.cpp
@@ -1,5 +1,4 @@
 #include "CesiumGeometry/clipTriangleAtAxisAlignedThreshold.h"
-#include "CesiumUtility/Profiler.h"
 
 namespace CesiumGeometry {
 void clipTriangleAtAxisAlignedThreshold(
